docs(core): add `@Transactional()` document
